refactor(painter): consolidate model→display-list compilation into painter/compile.rs
One named seam now owns the painter's v1-model reading. Mechanical
relocation — moved code is byte-identical, only imports changed.

Moved into the new src/painter/compile.rs:
- from painter/layer.rs: LayerList::{from_scene, from_node, flatten_node,
  build_render_commands, compute_clip_path}, the private flatten helpers
  (filter_visible_paints, filter_active_*, split_surface_effects,
  compute_rectangular_stroke_path, compute_non_overlapping_fill_path)
  and FlattenResult — as a second `impl LayerList` block, so
  LayerList::from_scene keeps its public signature (cache/scene.rs and
  runtime/scene.rs call sites unchanged).
- from painter/geometry.rs: build_shape(&Node, …), boolean_operation_path,
  boolean_operation_shape. geometry.rs keeps the model-agnostic
  PainterShape and merge_shapes only.

Also relocated LayerEffects from node/schema.rs to cg/fe.rs (it is a pure
composition of Fe* effect types with zero model coupling; schema.rs keeps
a pub use so every existing node::schema::LayerEffects path still works).
With that, layer.rs and painter.rs reference the model only through
crate::node::id::NodeId — the display list's sole model type.

Updated call sites: grida_editor/mode.rs and painter_debug_node.rs now
import build_shape/boolean_operation_shape from painter::compile.

Gates: cargo fmt -p grida clean; cargo check --workspace clean (0 warnings);
cargo clippy -p grida --no-deps clean; cargo test -p grida all green
(828 lib tests + all integration suites pass).

Part of the painter-narrowing seam gridaco/nothing#31 under the legacy
seam program gridaco/nothing#27.
