+++ crates/grida/src/painter/geometry.rs
@@ -1,20 +1,12 @@
 //! Shape building for rendering
 //!
-//! ## Pipeline Guarantees
-//!
-//! This module guarantees:
-//! - build_shape() requires resolved bounds from GeometryCache for all nodes
-//! - V2 nodes (with auto-sizing) ALWAYS use provided bounds
-//! - V1 nodes (fixed schema) use schema values (bounds parameter for future migration)
-//! - Missing bounds when accessed is a PANIC (pipeline bug)
+//! Model-agnostic shape abstraction ([`PainterShape`]) and path boolean
+//! merging. The model-facing shape construction (`build_shape` and the
+//! boolean-operation walks) lives in [`super::compile`] — the painter's
+//! single model seam (gridaco/nothing#31).
 
 use crate::cg::prelude::*;
-use crate::node::scene_graph::SceneGraph;
-use crate::node::schema::*;
 use crate::shape::*;
-use crate::{backends::skia as sk, cache::geometry::GeometryCache};
-use math2::rect::Rectangle;
-use math2::transform::AffineTransform;
 use skia_safe::{Path, RRect, Rect};
 
 /// Internal universal Painter's shape abstraction for optimized drawing
@@ -209,163 +201,6 @@ impl PainterShape {
     }
 }
 
-/// Build shape from node + resolved geometry
-///
-/// All dimensions come from bounds (resolved by GeometryCache).
-/// This ensures V2 auto-sized nodes and future migrations render correctly.
-pub fn build_shape(node: &Node, bounds: &Rectangle) -> PainterShape {
-    match node {
-        Node::Polygon(n) => {
-            let shape = n.to_shape();
-            PainterShape::from_shape(&shape)
-        }
-        Node::RegularPolygon(n) => {
-            let shape = n.to_shape();
-            PainterShape::from_shape(&shape)
-        }
-        Node::RegularStarPolygon(n) => {
-            let shape = n.to_shape();
-            PainterShape::from_shape(&shape)
-        }
-        Node::Line(n) => PainterShape::from_path(Path::line((0.0, 0.0), (n.size.width, 0.0))),
-        Node::Path(n) => {
-            if let Some(path) = Path::from_svg(&n.data) {
-                PainterShape::from_path(path)
-            } else {
-                // Fallback to empty rect if path is invalid
-                PainterShape::from_rect(Rect::new(0.0, 0.0, 0.0, 0.0))
-            }
-        }
-        Node::Vector(n) => {
-            let path = n.to_path();
-            PainterShape::from_path(path)
-        }
-        Node::Ellipse(n) => {
-            let shape = n.to_shape();
-            PainterShape::from_shape(&shape)
-        }
-        Node::Rectangle(n) => {
-            let shape = n.to_shape();
-            PainterShape::from_shape(&shape)
-        }
-        Node::Container(n) => {
-            // ALWAYS use resolved bounds from GeometryCache
-            let width = bounds.width;
-            let height = bounds.height;
-
-            let r = n.corner_radius;
-            if !r.is_zero() {
-                // Check if corner smoothing is enabled
-                if n.corner_smoothing.value() > 0.0 {
-                    let smooth = OrthogonalSmoothRRectShape {
-                        width,
-                        height,
-                        corner_radius: n.corner_radius,
-                        corner_smoothing: n.corner_smoothing,
-                    };
-                    PainterShape::from_path(build_orthogonal_smooth_rrect_path(&smooth))
-                } else {
-                    let rrect = build_rrect(&RRectShape {
-                        width,
-                        height,
-                        corner_radius: n.corner_radius,
-                    });
-                    PainterShape::from_rrect(rrect)
-                }
-            } else {
-                let rect = Rect::from_xywh(0.0, 0.0, width, height);
-                PainterShape::from_rect(rect)
-            }
-        }
-        Node::Image(n) => {
-            let r = n.corner_radius;
-            if !r.is_zero() {
-                // Check if corner smoothing is enabled
-                if n.corner_smoothing.value() > 0.0 {
-                    let smooth = OrthogonalSmoothRRectShape {
-                        width: n.size.width,
-                        height: n.size.height,
-                        corner_radius: r,
-                        corner_smoothing: n.corner_smoothing,
-                    };
-                    PainterShape::from_path(build_orthogonal_smooth_rrect_path(&smooth))
-                } else {
-                    let rrect = build_rrect(&n.to_own_shape());
-                    PainterShape::from_rrect(rrect)
-                }
-            } else {
-                let rect = Rect::from_xywh(0.0, 0.0, n.size.width, n.size.height);
-                PainterShape::from_rect(rect)
-            }
-        }
-        Node::Tray(n) => {
-            // Tray uses resolved bounds (like Container) with optional corner radius
-            let width = bounds.width;
-            let height = bounds.height;
-
-            let r = n.corner_radius;
-            if !r.is_zero() {
-                if n.corner_smoothing.value() > 0.0 {
-                    let smooth = OrthogonalSmoothRRectShape {
-                        width,
-                        height,
-                        corner_radius: n.corner_radius,
-                        corner_smoothing: n.corner_smoothing,
-                    };
-                    PainterShape::from_path(build_orthogonal_smooth_rrect_path(&smooth))
-                } else {
-                    let rrect = build_rrect(&RRectShape {
-                        width,
-                        height,
-                        corner_radius: n.corner_radius,
-                    });
-                    PainterShape::from_rrect(rrect)
-                }
-            } else {
-                let rect = Rect::from_xywh(0.0, 0.0, width, height);
-                PainterShape::from_rect(rect)
-            }
-        }
-        Node::Error(n) => {
-            let rect = Rect::from_xywh(0.0, 0.0, n.size.width, n.size.height);
-            PainterShape::from_rect(rect)
-        }
-        Node::MarkdownEmbed(n) => {
-            // Use resolved bounds for auto-sized markdown embeds
-            let width = bounds.width;
-            let height = bounds.height;
-            let r = n.corner_radius;
-            if !r.is_zero() {
-                let rrect = build_rrect(&RRectShape {
-                    width,
-                    height,
-                    corner_radius: n.corner_radius,
-                });
-                PainterShape::from_rrect(rrect)
-            } else {
-                let rect = Rect::from_xywh(0.0, 0.0, width, height);
-                PainterShape::from_rect(rect)
-            }
-        }
-        Node::HTMLEmbed(n) => {
-            let r = n.corner_radius;
-            if !r.is_zero() {
-                let rrect = build_rrect(&RRectShape {
-                    width: n.size.width,
-                    height: n.size.height,
-                    corner_radius: n.corner_radius,
-                });
-                PainterShape::from_rrect(rrect)
-            } else {
-                let rect = Rect::from_xywh(0.0, 0.0, n.size.width, n.size.height);
-                PainterShape::from_rect(rect)
-            }
-        }
-        // Non-shape nodes (Group, BooleanOperation, InitialContainer, TextSpan)
-        _ => PainterShape::from_rect(Rect::new(0.0, 0.0, 0.0, 0.0)),
-    }
-}
-
 /// Merges multiple shapes into a single path using boolean operations.
 ///
 /// This function takes a list of shapes and their corresponding boolean operations,
@@ -407,95 +242,4 @@ pub fn merge_shapes(shapes: &[(PainterShape, BooleanPathOperation)]) -> Path {
     }
 
     result
-}
-
-/// Compute the resulting path for a [`BooleanPathOperationNode`] in its local coordinate space.
-pub fn boolean_operation_path(
-    id: &NodeId,
-    node: &BooleanPathOperationNodeRec,
-    graph: &SceneGraph,
-    cache: &GeometryCache,
-) -> Option<Path> {
-    let world = cache
-        .get_world_transform(id)
-        .unwrap_or_else(AffineTransform::identity);
-    let inv = world.inverse().unwrap_or_else(AffineTransform::identity);
-
-    let mut shapes_with_ops = Vec::new();
-
-    let children = graph.get_children(id)?;
-    for (i, child_id) in children.iter().enumerate() {
-        if let Ok(child_node) = graph.get_node(child_id) {
-            let mut path = match child_node {
-                Node::BooleanOperation(child_bool) => {
-                    boolean_operation_path(child_id, child_bool, graph, cache)?
-                }
-                _ => {
-                    // Get bounds from geometry cache - guaranteed to exist
-                    let bounds = cache
-                        .get_world_bounds(child_id)
-                        .expect("Geometry must exist for all nodes");
-                    let intrinsic = match child_node {
-                        Node::Rectangle(n) => Node::Rectangle(n.clone()),
-                        Node::Ellipse(n) => Node::Ellipse(n.clone()),
-                        Node::Polygon(n) => Node::Polygon(n.clone()),
-                        Node::RegularPolygon(n) => Node::RegularPolygon(n.clone()),
-                        Node::RegularStarPolygon(n) => Node::RegularStarPolygon(n.clone()),
-                        Node::Line(n) => Node::Line(n.clone()),
-                        Node::Path(n) => Node::Path(n.clone()),
-                        Node::Vector(n) => Node::Vector(n.clone()),
-                        Node::Image(n) => Node::Image(n.clone()),
-                        Node::Container(n) => Node::Container(n.clone()),
-                        Node::Error(n) => Node::Error(n.clone()),
-                        // TODO: Support Group nodes as children of boolean operations.
-                        // Currently, boolean operations like `boolop(group(a, b), c)` are not supported.
-                        // When a Group is encountered, this function returns None, causing the boolean
-                        // operation to fail rendering.
-                        _ => return None, // Non-shape nodes (including Group)
-                    };
-                    build_shape(&intrinsic, &bounds).to_path()
-                }
-            };
-
-            let child_world = cache
-                .get_world_transform(child_id)
-                .unwrap_or_else(AffineTransform::identity);
-            let relative = inv.compose(&child_world);
-            path = path.make_transform(&sk::sk_matrix(relative.matrix));
-
-            let op = if i == 0 {
-                BooleanPathOperation::Union
-            } else {
-                node.op
-            };
-            shapes_with_ops.push((PainterShape::from_path(path), op));
-        }
-    }
-
-    if shapes_with_ops.is_empty() {
-        return None;
-    }
-
-    let path = merge_shapes(&shapes_with_ops);
-    let path = if let Some(r) = node.corner_radius {
-        if r > 0.0 {
-            build_corner_radius_path(&path, r)
-        } else {
-            path
-        }
-    } else {
-        path
-    };
-
-    Some(path)
-}
-
-/// Convenience wrapper around [`boolean_operation_path`] returning a [`PainterShape`].
-pub fn boolean_operation_shape(
-    id: &NodeId,
-    node: &BooleanPathOperationNodeRec,
-    graph: &SceneGraph,
-    cache: &GeometryCache,
-) -> Option<PainterShape> {
-    boolean_operation_path(id, node, graph, cache).map(PainterShape::from_path)
 }
